05_RAG/pdf_reader_streamlit/requirements.txt
# RAG Application Requirements
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.6
langchain-qdrant>=0.1.0
langchain-text-splitters>=0.0.1
qdrant-client>=1.7.0
pypdf>=4.0.0
openai>=1.12.0
python-dotenv>=1.0.0
